refactor(form-builder): one file per field
