+++ tools/bazar/presentation/javascripts/form-edit-template/fields/README
@@ -1,0 +1,5 @@
+export default {
+  attributes: {},
+  // disabledAttributes: [],
+  // renderInput(fieldData) {},
+}
