+++ tools/bazar/presentation/javascripts/form-edit-template/fields/collaborative_doc.js
@@ -1,0 +1,11 @@
+export default {
+  field: {
+    name: 'collaborative_doc',
+    attrs: { type: 'collaborative_doc' }
+  },
+  attributes: {},
+  // disabledAttributes: [],
+  renderInput(field) {
+    return { field: _t('BAZ_FORM_EDIT_COLLABORATIVE_DOC_FIELD') }
+  },
+}
