+++ tools/bazar/presentation/javascripts/form-edit-template/fields/commons/render-helper.js
@@ -1,5 +1,5 @@
 // function to render help for tabs and tabchange
-const templateHelper = {
+export default {
   cache: {},
   holders: {},
   ids: {},
