+++ tools/bazar/presentation/javascripts/form-edit-template/fields/file.js
@@ -1,0 +1,19 @@
+import { readConf, writeconf, semanticConf, defaultMapping } from './commons/attributes.js'
+
+export default {
+  attributes: {
+    readlabel: {
+      label: _t('BAZ_FORM_EDIT_FILE_READLABEL_LABEL'),
+      value: '',
+      placeholder: _t('BAZ_FILEFIELD_FILE')
+    },
+    maxsize: { label: _t('BAZ_FORM_EDIT_FILE_MAXSIZE_LABEL'), value: '' },
+    hint: { label: _t('BAZ_FORM_EDIT_HELP'), value: '' },
+    read: readConf,
+    write: writeconf,
+    semantic: semanticConf
+  },
+  // disabledAttributes: [],
+  attributesMapping: { ...defaultMapping, ...{ 3: 'maxsize', 6: 'readlabel' } },
+  // renderInput(fieldData) {},
+}
