+++ tools/bazar/presentation/javascripts/form-edit-template/fields/hidden.js
@@ -1,0 +1,6 @@
+export default {
+  attributes: {},
+  // disabledAttributes: [],
+  attributesMapping: { 0: 'type', 1: 'name', 5: 'value' },
+  // renderInput(fieldData) {},
+}
