+++ tools/bazar/presentation/javascripts/form-edit-template/fields/url.js
@@ -1,0 +1,19 @@
+import { readConf, writeconf, semanticConf, defaultMapping } from './commons/attributes.js'
+
+export default {
+  field: {
+    label: _t('BAZ_FORM_EDIT_URL_LABEL'),
+    name: 'url',
+    attrs: { type: 'url' },
+    icon: '<i class="fas fa-link"></i>'
+  },
+  attributes: {
+    read: readConf,
+    write: writeconf,
+    semantic: semanticConf
+  },
+  // disabledAttributes: [],
+  renderInput(fieldData) {
+    return { field: `<input type="url" placeholder="${fieldData.value || ''}"/>` }
+  }
+}
