+++ tools/bazar/templates/forms/_forms_form_js_include.twig
@@ -1,6 +1,3 @@
 {{ include_javascript('javascripts/vendor/jquery-ui-sortable/jquery-ui.min.js') }}
 {{ include_javascript('javascripts/vendor/formBuilder/form-builder.min.js') }}
-{{ include_javascript('tools/bazar/presentation/javascripts/form-edit-template-helper.js') }}
-{{ include_javascript('tools/bazar/presentation/javascripts/form-edit-template.js') }}
-{{ include_javascript('tools/bazar/presentation/javascripts/conditionschecking/form-edit-template.js') }}
-{{ include_javascript('tools/bazar/presentation/javascripts/calcfield/form-edit-template.js') }}
+{{ include_javascript('tools/bazar/presentation/javascripts/form-edit-template/form-edit-template.js', false, true) }}
