+++ tools/bazar/templates/inputs/conditions-checking.twig
@@ -1,2 +1,2 @@
-{{ include_javascript('tools/bazar/presentation/javascripts/conditionschecking/conditionschecking.js') }}
+{{ include_javascript('tools/bazar/presentation/javascripts/bazar-fields/conditionschecking.js') }}
 <div data-conditionschecking="{{ field.condition }}" {% if field.options.noclean %}data-noclean="true" {% endif %}style="display:none;">
